requirements.txt
beautifulsoup4==4.13.4
finvizfinance==1.1.0
pandas==2.2.3
plotly==5.24.1
pymongo==4.13.0
python-dotenv==1.0.0
Requests==2.32.3
scipy==1.11.4
streamlit==1.44.1
torch==2.1.0
tqdm==4.66.1
transformers==4.35.2
